+++ controlnext/tests/tests.py
@@ -3,6 +3,7 @@ import logging
 
 from django.test import TestCase
 
+import json
 import mock
 from pandas import Series
 import pytz
@@ -19,7 +20,7 @@ from controlnext.conf import settings
 from controlnext.demand_table import DemandTable
 from controlnext.fews_data import FewsJdbcDataSource
 from controlnext.utils import round_date, mktim
-from controlnext.wur_data import WURService
+from controlnext.wur_data import WURXMLService, WURJSONService
 from controlnext.tests.factories import GrowerInfoFactory, BasinFactory
 
 
@@ -154,18 +155,38 @@ def get_wur_service_mock_data(self):
     return DATA
 
 
-class WURServiceTest(TestCase):
+def get_wur_json_service_mock_data(self, url):
+    try:
+        from controlnext.tests.data.wur_response_from_json import DATA
+    except ImportError, info:
+        module_name = "controlnext.tests.data.wur_response_from_json"
+        msg = MOCK_DATA_IMPORT_ERROR % module_name
+        error_msg = "%s (%s)" % (msg, info)
+        logger.error(error_msg)
+        raise Exception(error_msg)
+    return DATA
+
+
+class WURServiceTests(TestCase):
 
     def setUp(self):
         now = datetime.datetime.now()
         _from = now - datetime.timedelta(days=2)
         to = now
-        self.service = WURService(_from, to)
+        self.xml_service = WURXMLService(_from, to)
 
     # TODO: mock suds response with wur_response.xml
-    @mock.patch('controlnext.wur_data.WURService.get_data',
+    @mock.patch('controlnext.wur_data.WURXMLService.get_data',
                 get_wur_service_mock_data)
     def test_get_data(self):
-        data = self.service.get_data()
+        data = self.xml_service.get_data()
         self.assertEqual(len(data), 48)  # 2 days
 
+    @mock.patch('controlnext.wur_data.WURJSONService.get_data_from_url',
+        get_wur_json_service_mock_data)
+    def test_json_service(self):
+        start_date = datetime.datetime(2013, 1, 1)
+        end_date = datetime.datetime(2013, 1, 2)
+        json_service = WURJSONService(start_date, end_date)
+        data = json_service.get_data()
+        self.assertGreater(len(data), 0)